mejoras en los Json

+++ Lab.Models/Grupo2/Frutas.cs
@@ -8,7 +8,7 @@ namespace Lab.Models.Grupo2
 {
     public class Frutas
     {
-        public int Cantidad { get; set; }
+        public float Cantidad { get; set; }
         public string TipodeFruta { get; set; }
         
     }
